transmute: refactor extension processing
Moving around the logic for third-party extension processing, where
individual third-party extensions will have their own implementation
file (instead of one big transmute file).
